refactor(schlib): split reader.rs into a reader/ module directory
Pure code motion (U3, completing the SchLib/PcbLib structural
unification): reader.rs becomes reader/{mod,parsers}.rs, mirroring
pcblib/reader/ — dispatch and shared helpers in mod.rs, the per-record
parse functions in parsers.rs. No file outside the reader module
changed; no behaviour change; gates green incl. the readability oracle
(0 regressions).
